:arrow_up: update xUnit project to RC version. Closes #844
- `project.json` migrated to `{APPNAME}.csproj`
- tests update

The content still has some example tests compared to dotnet/cli
version and the runner file is created together with default content
files

Thanks!
