--- templates/projects/xunit/Class1.cs
@@ -1,26 +1,0 @@
-using Xunit;
-
-namespace <%= namespace %>
-{
-    // see example explanation on xUnit.net website:
-    // https://xunit.github.io/docs/getting-started-dotnet-core.html
-    public class Class1
-    {
-        [Fact]
-        public void PassingTest()
-        {
-            Assert.Equal(4, Add(2, 2));
-        }
-
-        [Fact]
-        public void FailingTest()
-        {
-            Assert.Equal(5, Add(2, 2));
-        }
-
-        int Add(int x, int y)
-        {
-            return x + y;
-        }
-    }
-}
