--- templates/projects/xunit/project.json
@@ -1,26 +1,0 @@
-{
-    "version": "1.0.0-*",
-    "testRunner": "xunit",
-    "dependencies": {
-        "dotnet-test-xunit": "2.2.0-preview2-build1029",
-        "xunit": "2.2.0-beta2-build3300"
-    },
-    "frameworks": {
-        "netcoreapp1.0": {
-            "dependencies": {
-                "Microsoft.NETCore.App": {
-                    "type": "platform",
-                    "version": "1.0.1"
-                }
-            }
-        }
-    },
-    "buildOptions": {
-        "copyToOutput": {
-            "include": [ "xunit.runner.json" ]
-        }
-    },
-    "tooling": {
-        "defaultNamespace": "<%= namespace %>"
-    }
-}
